refactor: 重构路由导出和工具函数
- 在 `index.ts` 中更新路由导出，移除不必要的 `indx.ts` 文件，简化代码结构。
- 在 `app.ts` 中调整路由导入，确保使用最新的路由结构。
- 在 `getMimeType.ts` 中将 `getMimeType` 函数导出，提升可复用性。
- 在 `config/index.ts` 中移除不必要的导出，优化代码整洁性。
